backend/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
qrcode[pil]==7.4.2
pillow==11.1.0pip freeze > requirements.txt  